+++ frontend/main_window.py
@@ -1,6 +1,6 @@
 from frontend.file_selector import Menu
-from frontend.analiser import Analiser
-from data_processing.read_data import read_file
+from frontend.analiser.analiser import Analiser
+from data_reading.read_data import read_file
 
 
 class MainWindow:
